requirements.txt
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.10
vcrpy==4.2.0
wrapt==1.14.1
yarl==1.7.2
